requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
watchdog>=3.0.0
Pillow>=10.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
python-gphoto2>=2.5.0  # Optional: only needed for gphoto2 mode
rawpy>=0.19.0           # Optional: only needed if you want to process RAW→JPEG locally
